+++ lib/shrine/plugins/data_uri.rb
@@ -103,10 +103,7 @@ class Shrine
 
         # Returns contents of the file base64-encoded.
         def base64
-          io = storage.open(id)
-          content = io.read
-          io.close
-
+          content = open { |io| io.read }
           Base64.encode64(content).chomp
         end
       end